Move wgpu-canvas to renderer-common
