+++ ffi-run/Cargo.toml
@@ -14,6 +14,7 @@ uniffi = { version = "0.29.4", features = ["cli"]}
 pollster = { workspace = true }
 wgpu = { workspace = true}
 wgpu-canvas = { workspace = true }
+renderer-common = { workspace = true }
 log = { workspace = true }
 osm = { workspace = true }
