+++ map/src/camera.rs
@@ -5,7 +5,7 @@ use glam::DVec2;
 use glam::DVec3;
 use glam::Vec3Swizzles;
 use std::f64::consts::PI;
-use wgpu_canvas::LIGHT_POS;
+use renderer_common::LIGHT_POS;
 
 pub struct Camera {
     pub eye: DVec3,
