+++ renderer/src/draw_commands/mesh2d_draw_command.rs
@@ -3,7 +3,7 @@ use crate::draw_commands::DrawCommand;
 use crate::global_context::GlobalContext;
 use crate::mesh::mesh::{Mesh, StyledRange};
 use crate::mesh_layers::layers::Layers;
-use wgpu_canvas::render_modifier::SpatialData;
+use renderer_common::render_modifier::SpatialData;
 use crate::vertex_attrs::ShapeVertex;
 use lyon::tessellation::VertexBuffers;
 use std::mem;
