+++ renderer/src/mesh/positioned_mesh.rs
@@ -2,7 +2,7 @@ use crate::global_context::GlobalContext;
 use crate::mesh::mesh::Mesh;
 use crate::mesh::mesh_instance_input::MeshInstanceInput;
 use crate::mesh::InstanceBuffer;
-use wgpu_canvas::render_modifier::SpatialData;
+use renderer_common::render_modifier::SpatialData;
 use crate::pipelines::IndirectInstancesLayout;
 use crate::utils::ReceiverExt;
 use glam::DVec3;
