+++ renderer/src/mesh_layers/screen_shape_layer.rs
@@ -6,7 +6,7 @@ use crate::mesh::mesh::Mesh;
 use crate::mesh::mesh_instance_input::MeshInstanceInput;
 use crate::mesh::InstanceBuffer;
 use crate::mesh_layers::BaseMeshLayer;
-use wgpu_canvas::render_modifier::SpatialData;
+use renderer_common::render_modifier::SpatialData;
 use crate::pipelines::RenderPipeline;
 use crate::view_projection::ViewProjection;
 use geo_types::point;
