+++ renderer/src/mesh_layers/text_mesh_layer.rs
@@ -1,7 +1,7 @@
-use wgpu_canvas::geometry_data::{LineData, TextData};
+use renderer_common::geometry_data::{LineData, TextData};
 use crate::global_context::GlobalContext;
 use crate::mesh_layers::BaseMeshLayer;
-use wgpu_canvas::render_modifier::SpatialData;
+use renderer_common::render_modifier::SpatialData;
 use crate::pipelines::RenderPipeline;
 use crate::text::text_renderer::TextRenderer;
 use wgpu::{CommandEncoder, RenderPass};
