+++ renderer/src/renderer_api.rs
@@ -2,11 +2,11 @@ use crate::canvas_api::GpuCanvasApi;
 use crate::messages::RendererApiMsg;
 use std::collections::HashSet;
 use std::sync::mpsc::Sender;
-use wgpu_canvas::render_group::RenderGroup;
-use wgpu_canvas::render_modifier::SpatialData;
-use wgpu_canvas::render_style::RenderStyle;
-use wgpu_canvas::style_id::StyleId;
-use wgpu_canvas::RendererApi;
+use renderer_common::render_group::RenderGroup;
+use renderer_common::render_modifier::SpatialData;
+use renderer_common::render_style::RenderStyle;
+use renderer_common::style_id::StyleId;
+use renderer_common::RendererApi;
 
 pub struct GpuRendererApi {
     renderer_api_tx: Sender<RendererApiMsg>,
