+++ winit-run/src/main.rs
@@ -19,7 +19,7 @@ use wgpu::TextureUsages;
 use wgpu::{Features, Limits};
 use renderer::GpuRenderer;
 use renderer::wgpu_canvas::DefaultWgpuCanvas;
-use wgpu_canvas::{feature_layer_tags, PreviewType, PREVIEW_TYPE, SHADOWS_ENABLED, SHADOWS_TEX_SIZE, SSAO_ENABLED};
+use renderer_common::{feature_layer_tags, PreviewType, PREVIEW_TYPE, SHADOWS_ENABLED, SHADOWS_TEX_SIZE, SSAO_ENABLED};
 
 slint::include_modules!();
